refactor one class per exchange
