+++ src/book.py
@@ -37,6 +37,16 @@ import misc
 import transaction as tr
 from core import kraken_asset_map
 from database import set_price_db
+from exchanges.base import ExchangeReader
+from exchanges.binance import BinanceReader
+from exchanges.bitget_api import BitgetApiReader
+from exchanges.bitpanda import BitpandaReader
+from exchanges.bitunix import BitunixReader
+from exchanges.coinbase import CoinbaseReader
+from exchanges.coinbase_pro import CoinbaseProReader
+from exchanges.custom_eur import CustomEurReader
+from exchanges.kraken import KrakenReader
+from exchanges.pionex import PionexReader
 from price_data import PriceData
 
 log = log_config.getLogger(__name__)
@@ -48,6 +58,36 @@ class MissingOperation(NamedTuple):
 
     def repr(self) -> str:
         return f"- {self.platform}: {self.operation}"
+
+
+def create_exchange_reader(exchange_name: str) -> Optional[ExchangeReader]:
+    """Create an exchange reader instance based on the exchange name."""
+    reader_map = {
+        "binance": BinanceReader,
+        "binance_v2": BinanceReader,
+        "binance_v3": BinanceReader,
+        "coinbase": CoinbaseReader,
+        "coinbase_v2": CoinbaseReader,
+        "coinbase_v3": CoinbaseReader,
+        "coinbase_v4": CoinbaseReader,
+        "coinbase_pro": CoinbaseProReader,
+        "kraken_ledgers_old": KrakenReader,
+        "kraken_ledgers": KrakenReader,
+        "kraken_trades": KrakenReader,
+        "bitpanda_pro_trades": BitpandaReader,
+        "bitpanda": BitpandaReader,
+        "bitunix": BitunixReader,
+        "pionex_deposit_withdraw": PionexReader,
+        "pionex_trading": PionexReader,
+        "pionex_staking": PionexReader,
+        "pionex_others": PionexReader,
+        "custom_eur": CustomEurReader,
+    }
+
+    reader_class = reader_map.get(exchange_name)
+    if reader_class:
+        return reader_class()
+    return None
 
 
 class Book:
@@ -136,2442 +176,6 @@ class Book:
             if op is not None:
                 self._append_operation(op)
 
-    def _read_binance(self, file_path: Path, version: int = 1) -> None:
-        platform = "binance"
-        operation_mapping = {
-            "Distribution": "Airdrop",
-            "Cash Voucher distribution": "Airdrop",
-            "Cashback Voucher": "Airdrop",
-            "Rewards Distribution": "Airdrop",
-            "Simple Earn Flexible Airdrop": "Airdrop",
-            "Airdrop Assets": "Airdrop",
-            "Crypto Box": "Airdrop",
-            "Launchpool Airdrop": "Airdrop",
-            "Megadrop Rewards": "Airdrop",
-            "HODLer Airdrops Distribution": "Airdrop",
-            "Token Swap - Distribution": "Airdrop",
-            "Launchpool Airdrop - System Distribution": "Airdrop",
-            #
-            "Savings Interest": "CoinLendInterest",
-            "Savings purchase": "CoinLend",
-            "Savings Principal redemption": "CoinLendEnd",
-            "Savings distribution": "CoinLendInterest",
-            "Simple Earn Flexible Subscription": "CoinLend",
-            "Simple Earn Flexible Redemption": "CoinLendEnd",
-            "Simple Earn Flexible Interest": "CoinLendInterest",
-            "Simple Earn Locked Subscription": "CoinLend",
-            "Simple Earn Locked Redemption": "CoinLendEnd",
-            "Simple Earn Locked Rewards": "CoinLendInterest",
-            "Savings Distribution": "CoinLendInterest",
-            #
-            "BNB Vault Rewards": "CoinLendInterest",
-            "Launchpool Earnings Withdrawal": "CoinLendInterest",
-            #
-            "Commission History": "Commission",
-            "Commission Fee Shared With You": "Commission",
-            "Referrer rebates": "Commission",
-            "Referral Kickback": "Commission",
-            "Commission Rebate": "Commission",
-            # DeFi yield farming
-            "Liquid Swap add": "CoinLend",
-            "Liquid Swap remove": "CoinLendEnd",
-            "Liquid Swap rewards": "CoinLendInterest",
-            "Launchpool Interest": "CoinLendInterest",
-            #
-            "Super BNB Mining": "StakingInterest",
-            "POS savings interest": "StakingInterest",
-            "POS savings purchase": "Staking",
-            "POS savings redemption": "StakingEnd",
-            "ETH 2.0 Staking Rewards": "StakingInterest",
-            "Staking Purchase": "Staking",
-            "Staking Rewards": "StakingInterest",
-            "Staking Redemption": "StakingEnd",
-            #
-            "Fiat Deposit": "Deposit",
-            "Fiat Withdraw": "Withdrawal",
-            "Withdraw": "Withdrawal",
-            #
-            "Transaction Buy": "Buy",
-            "Transaction Spend": "Sell",
-            "Transaction Revenue": "Buy",
-            "Transaction Sold": "Sell",
-            "Transaction Fee": "Fee",
-            "Asset Recovery": "Sell",
-        }
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for rowlist in reader:
-                if version == 1:
-                    _utc_time, account, operation, coin, _change, remark = rowlist
-                elif version in (2, 3):
-                    (
-                        _,
-                        _utc_time,
-                        account,
-                        operation,
-                        coin,
-                        _change,
-                        remark,
-                    ) = rowlist
-                else:
-                    log.error("File version not Supported " + str(file_path))
-                    raise NotImplementedError
-
-                row = reader.line_num
-
-                # Parse data.
-                if version in (1, 2):
-                    utc_time = datetime.datetime.strptime(
-                        _utc_time, "%Y-%m-%d %H:%M:%S"
-                    )
-                elif version == 3:
-                    utc_time = datetime.datetime.strptime(
-                        _utc_time, "%y-%m-%d %H:%M:%S"
-                    )
-                else:
-                    log.error("File version not Supported " + str(file_path))
-                    raise NotImplementedError
-
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-                change = misc.force_decimal(_change)
-                operation = operation_mapping.get(operation, operation)
-                if operation in (
-                    "The Easiest Way to Trade",
-                    "Small assets exchange BNB",
-                    "Small Assets Exchange BNB",
-                    "Transaction Related",
-                    "Large OTC trading",
-                    "Sell",
-                    "Buy",
-                    "Binance Convert",
-                ):
-                    operation = "Sell" if change < 0 else "Buy"
-
-                if operation == "Liquid Swap add/sell":
-                    operation = "CoinLendEnd" if change < 0 else "CoinLend"
-
-                if operation == "Commission" and account != "Spot":
-                    # All comissions will be handled the same way.
-                    # As of now, only Spot Binance Operations are supported,
-                    # so we have to change the account type to Spot.
-                    account = "Spot"
-
-                if (
-                    account in ("Spot", "P2P")
-                    and operation
-                    in (
-                        "transfer_in",
-                        "transfer_out",
-                    )
-                    or (
-                        account in ("Spot", "Funding")
-                        and operation == "Transfer Between Main and Funding Wallet"
-                    )
-                ):
-                    # Ignore transfers
-                    continue
-
-                change = abs(change)
-
-                # Validate data.
-                supported_account_types = ("Spot", "Savings", "Earn", "Funding")
-                assert account in supported_account_types, (
-                    f"Other types than {supported_account_types} are currently "
-                    f"not supported.  Given account type is `{account}`. "
-                    "Please create an Issue or PR."
-                )
-                assert operation
-                assert coin
-                assert change
-
-                if remark:
-                    # Ignore default remarks
-                    if remark in (
-                        "Withdraw fee is included",
-                        "Binance Earn",
-                        "Binance Pay",
-                        "Binance Launchpool",
-                    ) or remark.endswith(" to BNB"):
-                        remark = ""
-
-                    # Do not warn for specific remarks
-                    elif remark.startswith("Korrekturbuchung."):
-                        pass
-
-                    # Warn on other binance remarks, becuase all remarks should be some
-                    # unnecessary default text which we'd like to ignore
-                    else:
-                        log.warning(
-                            "I may have missed a remark in %s:%i: `%s`.",
-                            file_path,
-                            row,
-                            remark,
-                        )
-
-                self.append_operation(
-                    operation, utc_time, platform, change, coin, row, file_path, remark
-                )
-
-    def _read_binance_v2(self, file_path: Path) -> None:
-        self._read_binance(file_path=file_path, version=2)
-
-    def _read_binance_v3(self, file_path: Path) -> None:
-        self._read_binance(file_path=file_path, version=3)
-
-    def _read_coinbase(self, file_path: Path, version: int = 1) -> None:
-        platform = "coinbase"
-        operation_mapping = {
-            "Receive": "Deposit",
-            "Send": "Withdrawal",
-            "Coinbase Earn": "Buy",
-            "Learning Reward": "Buy",
-            "Rewards Income": "Staking",
-        }
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            try:
-                if version == 4:
-                    assert next(reader) == []
-                    assert next(reader) == ["Transactions"]
-                    assert next(reader)  # user row
-                else:
-                    assert next(reader)  # header line
-                    assert next(reader) == []
-                    assert next(reader) == []
-                    assert next(reader) == []
-                    assert next(reader) == ["Transactions"]
-                    assert next(reader)  # user row
-                    assert next(reader) == []
-
-                fields = next(reader)
-                num_columns = len(fields)
-                # Coinbase export format from 2023/2024 and ongoing
-                if num_columns == 11:
-                    assert version == 4
-                    assert fields == [
-                        "ID",
-                        "Timestamp",
-                        "Transaction Type",
-                        "Asset",
-                        "Quantity Transacted",
-                        "Price Currency",
-                        "Price at Transaction",
-                        "Subtotal",
-                        "Total (inclusive of fees and/or spread)",
-                        "Fees and/or Spread",
-                        "Notes",
-                    ]
-                # Coinbase export format from late 2021 until 2023/2024
-                elif num_columns == 10:
-                    assert fields == [
-                        "Timestamp",
-                        "Transaction Type",
-                        "Asset",
-                        "Quantity Transacted",
-                        "Spot Price Currency",
-                        "Spot Price at Transaction",
-                        "Subtotal",
-                        "Total (inclusive of fees)",
-                        "Fees",
-                        "Notes",
-                    ] or fields == [
-                        "Timestamp",
-                        "Transaction Type",
-                        "Asset",
-                        "Quantity Transacted",
-                        "Spot Price Currency",
-                        "Spot Price at Transaction",
-                        "Subtotal",
-                        "Total (inclusive of fees and/or spread)",
-                        "Fees and/or Spread",
-                        "Notes",
-                    ]
-                # Coinbase export format from mid 2021 and before
-                elif num_columns == 9:
-                    assert fields == [
-                        "Timestamp",
-                        "Transaction Type",
-                        "Asset",
-                        "Quantity Transacted",
-                        "EUR Spot Price at Transaction",
-                        "EUR Subtotal",
-                        "EUR Total (inclusive of fees)",
-                        "EUR Fees",
-                        "Notes",
-                    ]
-                else:
-                    raise RuntimeError(
-                        "Unknown Coinbase format: "
-                        "Number of rows do not match known versions: "
-                        f"{file_path}."
-                    )
-            except AssertionError as e:
-                msg = (
-                    "Unable to read coinbase file: Malformed header. "
-                    f"Skipping {file_path}."
-                )
-                e.args += (msg,)
-                log.exception(e)
-                return
-
-            for columns in reader:
-
-                # Coinbase export format from 2023/2024 and ongoing
-                if num_columns == 11:
-                    (
-                        _id,
-                        _utc_time,
-                        operation,
-                        coin,
-                        _change,
-                        _currency_spot,
-                        _eur_spot,
-                        _eur_subtotal,
-                        _eur_total,
-                        _eur_fee,
-                        remark,
-                    ) = columns
-                    _eur_spot = _eur_spot.replace("€", "")
-                    _eur_subtotal = _eur_subtotal.replace("€", "")
-                    _eur_total = _eur_total.replace("€", "")
-                    _eur_fee = _eur_fee.replace("€", "")
-
-                # Coinbase export format from late 2021 until 2023/2024
-                elif num_columns == 10:
-                    (
-                        _utc_time,
-                        operation,
-                        coin,
-                        _change,
-                        _currency_spot,
-                        _eur_spot,
-                        _eur_subtotal,
-                        _eur_total,
-                        _eur_fee,
-                        remark,
-                    ) = columns
-
-                # Coinbase export format from mid 2021 and before
-                elif num_columns == 9:
-                    (
-                        _utc_time,
-                        operation,
-                        coin,
-                        _change,
-                        _eur_spot,  # Rounded price from CSV, unused
-                        _eur_subtotal,  # Cost without fees
-                        _eur_total,
-                        _eur_fee,
-                        remark,
-                    ) = columns
-                    _currency_spot = "EUR"
-
-                row = reader.line_num
-
-                # Parse data.
-                if version == 4:
-                    utc_time = datetime.datetime.strptime(
-                        _utc_time, "%Y-%m-%d %H:%M:%S UTC"
-                    )
-                else:
-                    utc_time = datetime.datetime.strptime(
-                        _utc_time, "%Y-%m-%dT%H:%M:%SZ"
-                    )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-                operation = operation_mapping.get(operation, operation)
-                change = misc.force_decimal(_change)
-                # `eur_subtotal` and `eur_fee` are None for withdrawals.
-                eur_subtotal = misc.xdecimal(_eur_subtotal)
-                if version == 4:
-                    change = abs(change)
-                    eur_subtotal = abs(eur_subtotal) if eur_subtotal else None
-                if eur_subtotal is None:
-                    # Cost without fees from CSV is missing. This can happen for
-                    # old transactions (<2018), event though something was bought.
-                    # Calculate the `eur_subtotal` from `eur_spot`.
-                    if eur_spot := misc.xdecimal(_eur_spot):
-                        eur_subtotal = eur_spot * change
-                eur_fee = misc.xdecimal(_eur_fee)
-
-                # Validate data.
-                assert operation
-                assert coin
-                assert change
-                assert _currency_spot == "EUR"
-
-                # Calculated price
-                if eur_subtotal:
-                    assert isinstance(eur_subtotal, decimal.Decimal)
-                    price_calc = eur_subtotal / change
-                    # Save price in our local database for later.
-                    set_price_db(platform, coin, "EUR", utc_time, price_calc)
-
-                if operation == "Convert":
-                    # Parse change + coin from remark, which is
-                    # in format "Converted 0,123 ETH to 0,456 BTC".
-                    match = re.match(
-                        r"^Converted [0-9,\.]+ [A-Z]+ to "
-                        r"(?P<change>[0-9,\.]+) (?P<coin>[A-Z]+)$",
-                        remark,
-                    )
-                    assert match
-
-                    _convert_change = match.group("change").replace(",", ".")
-                    convert_change = misc.force_decimal(_convert_change)
-                    convert_coin = match.group("coin")
-
-                    eur_total = misc.force_decimal(_eur_total)
-                    if version == 4:
-                        eur_total = abs(eur_total)
-                    convert_eur_spot = eur_total / convert_change
-
-                    self.append_operation(
-                        "Sell", utc_time, platform, change, coin, row, file_path
-                    )
-                    self.append_operation(
-                        "Buy",
-                        utc_time,
-                        platform,
-                        convert_change,
-                        convert_coin,
-                        row,
-                        file_path,
-                    )
-
-                    # Save convert price in local database, too.
-                    set_price_db(
-                        platform, convert_coin, "EUR", utc_time, convert_eur_spot
-                    )
-                else:
-                    # Add operation normally to the list.
-                    self.append_operation(
-                        operation, utc_time, platform, change, coin, row, file_path
-                    )
-
-                    # If it's a sell, add the corresponding buy to complement
-                    # the trading pair.
-                    if operation == "Sell":
-                        assert isinstance(eur_subtotal, decimal.Decimal)
-                        self.append_operation(
-                            "Buy",
-                            utc_time,
-                            platform,
-                            eur_subtotal,
-                            "EUR",
-                            row,
-                            file_path,
-                        )
-                    # If it's a buy, add the corresponding sell to complement
-                    # the trading pair.
-                    elif operation == "Buy":
-                        assert isinstance(eur_subtotal, decimal.Decimal)
-                        self.append_operation(
-                            "Sell",
-                            utc_time,
-                            platform,
-                            eur_subtotal,
-                            "EUR",
-                            row,
-                            file_path,
-                        )
-
-                # Add paid fees to the list.
-                if eur_fee:
-                    assert isinstance(eur_fee, decimal.Decimal)
-                    self.append_operation(
-                        "Fee", utc_time, platform, eur_fee, "EUR", row, file_path
-                    )
-
-    def _read_coinbase_v2(self, file_path: Path) -> None:
-        self._read_coinbase(file_path=file_path, version=2)
-
-    def _read_coinbase_v3(self, file_path: Path) -> None:
-        self._read_coinbase(file_path=file_path, version=3)
-
-    def _read_coinbase_v4(self, file_path: Path) -> None:
-        self._read_coinbase(file_path=file_path, version=4)
-
-    def _read_coinbase_pro(self, file_path: Path) -> None:
-        platform = "coinbase_pro"
-        operation_mapping = {
-            "BUY": "Buy",
-            "SELL": "Sell",
-        }
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for (
-                portfolio,
-                trade_id,
-                product,
-                operation,
-                _utc_time,
-                _size,
-                size_unit,
-                _price,
-                _fee,
-                total,
-                price_fee_total_unit,
-            ) in reader:
-                row = reader.line_num
-
-                # Parse data.
-                utc_time = datetime.datetime.strptime(
-                    _utc_time, "%Y-%m-%dT%H:%M:%S.%fZ"
-                )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-                operation = operation_mapping.get(operation, operation)
-                size = misc.force_decimal(_size)
-                price = misc.force_decimal(_price)
-                fee = misc.xdecimal(_fee)
-                total_price = size * price
-
-                # Unused variables.
-                del portfolio
-                del trade_id
-                del product
-                del total
-
-                # Validate data.
-                assert operation
-                assert size
-                assert size_unit
-                assert price_fee_total_unit
-
-                self.append_operation(
-                    operation, utc_time, platform, size, size_unit, row, file_path
-                )
-
-                if operation == "Sell":
-                    self.append_operation(
-                        "Buy",
-                        utc_time,
-                        platform,
-                        total_price,
-                        price_fee_total_unit,
-                        row,
-                        file_path,
-                    )
-                elif operation == "Buy":
-                    self.append_operation(
-                        "Sell",
-                        utc_time,
-                        platform,
-                        total_price,
-                        price_fee_total_unit,
-                        row,
-                        file_path,
-                    )
-                if fee:
-                    self.append_operation(
-                        "Fee",
-                        utc_time,
-                        platform,
-                        fee,
-                        price_fee_total_unit,
-                        row,
-                        file_path,
-                    )
-
-    def _read_kraken_trades(self, file_path: Path) -> None:
-        log.error(
-            f"{file_path.name}: "
-            "Looks like this is a Kraken 'Trades' history, "
-            "but we need the 'Ledgers' history. "
-            "(See: Wiki - Exchange Kraken)"
-        )
-
-    def _read_kraken_ledgers(self, file_path: Path) -> None:
-        fee_sign_of_file: Optional[bool] = None
-
-        platform = "kraken"
-        operation_mapping = {
-            "spend": "Sell",  # Sell ordered via 'Buy Crypto' button
-            "receive": "Buy",  # Buy ordered via 'Buy Crypto' button
-            "reward": "StakingInterest",
-            "staking": "StakingInterest",
-            "deposit": "Deposit",
-            "withdrawal": "Withdrawal",
-        }
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for columns in reader:
-
-                num_columns = len(columns)
-                # Kraken ledgers export format from October 2020 and ongoing
-                if num_columns == 10:
-                    (
-                        txid,
-                        refid,
-                        _utc_time,
-                        _type,
-                        subtype,
-                        aclass,
-                        _asset,
-                        _amount,
-                        _fee,
-                        balance,
-                    ) = columns
-
-                # Kraken ledgers export format from September 2020 and before
-                elif num_columns == 9:
-                    (
-                        txid,
-                        refid,
-                        _utc_time,
-                        _type,
-                        aclass,
-                        _asset,
-                        _amount,
-                        _fee,
-                        balance,
-                    ) = columns
-                else:
-                    log.error(
-                        f"{file_path}: Unknown Kraken ledgers format: "
-                        "Number of rows do not match known versions."
-                    )
-                    raise RuntimeError
-
-                row = reader.line_num
-
-                # Parse data.
-                utc_time = datetime.datetime.strptime(_utc_time, "%Y-%m-%d %H:%M:%S")
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-                change = misc.force_decimal(_amount)
-                # remove the appended .S for staked assets
-                _asset = _asset.removesuffix(".S")
-                coin = kraken_asset_map.get(_asset, _asset)
-                fee = misc.force_decimal(_fee)
-                # An older implementation expected always positive fees
-                # It seems that newer ledger files can have negative fee
-                # values instead.
-                if fee != 0:
-                    # As soon as the first fee!=0 appears, check whether the
-                    # fees are positive or negative. All fees in the file
-                    # should have the same sign.
-                    if fee_sign_of_file is None:
-                        fee_sign_of_file = fee < 0
-                    # Adjust the fee sign so that fees are always positive.
-                    if fee_sign_of_file is True:
-                        fee *= -1
-                    if fee < 0:
-                        log.error(
-                            f"{file_path} row {row}: Unexpected fee sign. "
-                            "All fees should have the same sign. "
-                            "Please create an Issue or PR."
-                        )
-                        raise RuntimeError
-
-                operation = operation_mapping.get(_type)
-                if operation is None:
-                    if _type == "trade":
-                        operation = "Sell" if change < 0 else "Buy"
-                    elif _type in ["margin trade", "rollover", "settled", "margin"]:
-                        log.error(
-                            f"{file_path} row {row}: Margin trading is currently not "
-                            "supported. Please create an Issue or PR."
-                        )
-                        raise RuntimeError
-                    elif _type == "transfer":
-                        if num_columns == 9:
-                            # for backwards compatibility assume Airdrop for staking
-                            log.warning(
-                                f"{file_path} row {row}: Staking is not supported for"
-                                "old Kraken ledger formats. "
-                                "Please create an Issue or PR."
-                            )
-                            operation = "Airdrop"
-                        elif subtype == "stakingfromspot":
-                            operation = "Staking"
-                        elif subtype == "stakingtospot":
-                            operation = "StakingEnd"
-                        elif subtype in ["spottostaking", "spotfromstaking"]:
-                            # duplicate entries for staking actions
-                            continue
-                        else:
-                            log.error(
-                                f"{file_path} row {row}: Order subtype '{subtype}' is "
-                                "currently not supported. Please create an Issue or PR."
-                            )
-                            raise RuntimeError
-                    else:
-                        log.error(
-                            f"{file_path} row {row}: Other order type '{_type}' is "
-                            "currently not supported. Please create an Issue or PR."
-                        )
-                        raise RuntimeError
-                change = abs(change)
-
-                # Validate data.
-                assert operation
-                assert coin
-                assert change
-
-                # Skip duplicate entries for deposits / withdrawals and additional
-                # deposit / withdrawal lines for staking / unstaking / staking reward
-                # actions.
-                # The second deposit and the first withdrawal need to be considered,
-                # since these are the points in time where the user actually has the
-                # assets at their disposal. The first deposit and second withdrawal are
-                # in the public trade history and are skipped.
-                # For staking / unstaking / staking reward actions, deposits /
-                # withdrawals only occur once and will be ignored.
-                # The "appended" flag stores if an operation for a given refid has
-                # already been appended to the operations list:
-                # == None: Initial value (first occurrence)
-                # == False: No operation has been appended (second occurrence)
-                # == True: Operation has already been appended, this should not happen
-                if operation in ["Deposit", "Withdrawal"]:
-                    # First, create the operations
-                    op = self._create_operation(
-                        operation, utc_time, platform, change, coin, row, file_path
-                    )
-                    op_fee = None
-                    if fee != 0:
-                        op_fee = self._create_operation(
-                            "Fee", utc_time, platform, fee, coin, row, file_path
-                        )
-                    if (op is None) or (fee != 0 and op_fee is None):
-                        # Ignore on this run - operation could not be created
-                        # This might lead to unexpected errors while parsing the
-                        # rest of the file...
-                        # It'll be fixed, when the _missing_operation_mappings
-                        # aren't missing.
-                        pass
-                    # If this is the first occurrence, set the "appended" flag to false
-                    # and don't append the operation to the list. Instead, store the
-                    # data for verifying or appending it later.
-                    elif self.kraken_held_ops[refid]["appended"] is None:
-                        self.kraken_held_ops[refid]["appended"] = False
-                        self.kraken_held_ops[refid]["operation"] = op
-                        self.kraken_held_ops[refid]["operation_fee"] = op_fee
-                    # If this is the second occurrence, append a new operation, set the
-                    # "appended" flag to True and assert that the data of this operation
-                    # agrees with the data of the first occurrence.
-                    elif self.kraken_held_ops[refid]["appended"] is False:
-                        self.kraken_held_ops[refid]["appended"] = True
-                        try:
-                            # Make sure, that the found operations with the
-                            # same refid  have the same operation type, amount
-                            # of change and same coin.
-                            assert isinstance(
-                                op, type(self.kraken_held_ops[refid]["operation"])
-                            ), (
-                                "operation "
-                                f"({op.type_name} != "
-                                f'{self.kraken_held_ops[refid]["operation"].type_name})'
-                            )
-                            assert (
-                                op.change
-                                == self.kraken_held_ops[refid]["operation"].change
-                            ), (
-                                "change "
-                                f"({op.change} != "
-                                f'{self.kraken_held_ops[refid]["operation"].change})'
-                            )
-                            assert (
-                                op.coin == self.kraken_held_ops[refid]["operation"].coin
-                            ), (
-                                "coin "
-                                f"({op.coin} != "
-                                f'{self.kraken_held_ops[refid]["operation"].coin})'
-                            )
-                        except AssertionError as e:
-                            # Row is internally saved as list[int].
-                            first_row = self.kraken_held_ops[refid]["operation"].line[0]
-                            log.error(
-                                "Two internal kraken operations matched by the "
-                                f"same {refid=} don't have the same {e}.\n"
-                                "CoinTaxman expects, that these two operations "
-                                "have the same type of operation, amount of "
-                                "change and the same coin.\n"
-                                f"See {file_path} in row {first_row} and "
-                                f"{row}.\n"
-                                "Please create an Issue or PR."
-                            )
-                            raise RuntimeError
-                        # For deposits, this is all we need to do before appending the
-                        # operation. For withdrawals, we need to append the first
-                        # withdrawal as soon as the second withdrawal occurs. Therefore,
-                        # overwrite the operation with the stored first withdrawal.
-                        if operation == "Withdrawal":
-                            op = self.kraken_held_ops[refid]["operation"]
-                            op_fee = self.kraken_held_ops[refid]["operation_fee"]
-                        # Finally, append the operations and delete the stored
-                        # operations to reduce memory consumption
-                        self._append_operation(op)
-                        if op_fee:
-                            self._append_operation(op_fee)
-                        del self.kraken_held_ops[refid]["operation"]
-                        del self.kraken_held_ops[refid]["operation_fee"]
-                    # If an operation with the same refid has been already appended,
-                    # this is the third occurrence. Throw an error if this happens.
-                    elif self.kraken_held_ops[refid]["appended"] is True:
-                        log.error(
-                            f"{file_path} row {row}: More than two entries with refid "
-                            f"{refid} should not exist ({operation}). "
-                            "Please create an Issue or PR."
-                        )
-                        raise RuntimeError
-                    # This should never happen
-                    else:
-                        log.error(
-                            f"{file_path} row {row}: Unknown value for appended "
-                            f"operation flag {self.kraken_held_ops[refid]['appended']}."
-                            "Please create an Issue or PR."
-                        )
-                        raise TypeError
-
-                # for all other operation types
-                else:
-                    self.append_operation(
-                        operation, utc_time, platform, change, coin, row, file_path
-                    )
-                    if fee != 0:
-                        self.append_operation(
-                            "Fee", utc_time, platform, fee, coin, row, file_path
-                        )
-                    if operation == "StakingInterest":
-                        # For Kraken, the rewarded coins are added to the staked
-                        # portfolio. TODO (for MULTI_DEPOT only): Directly add the
-                        # rewarded coins to the staking depot (not like here with the
-                        # detour of adding it to spot and then staking the same amount)
-                        self.append_operation(
-                            "Staking", utc_time, platform, change, coin, row, file_path
-                        )
-
-    def _read_kraken_ledgers_old(self, file_path: Path) -> None:
-        self._read_kraken_ledgers(file_path)
-
-    def _read_bitpanda_pro_trades(self, file_path: Path) -> None:
-        """Reads a trade statement from Bitpanda Pro.
-
-        Args:
-            file_path (Path): Path to Bitpanda trade history.
-        """
-
-        platform = "bitpanda_pro"
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # skip header
-            next(reader)
-            line = next(reader)
-
-            transaction_file_warn = (
-                f"{file_path} looks like a Bitpanda transaction file."
-                " Skipping. Please download the trade history instead."
-            )
-
-            # for transactions, it's currently written "id" (small)
-            if line[0].startswith("Account id :"):
-                log.warning(transaction_file_warn)
-                return
-
-            assert line[0].startswith("Account ID:")
-            line = next(reader)
-            # empty line - still keep this check in case Bitpanda changes the
-            # transaction file to match the trade header (casing)
-            if not line:
-                log.warning(transaction_file_warn)
-                return
-
-            elif line[0] != "Bitpanda Pro trade history":
-                log.warning(
-                    f"{file_path} doesn't look like a Bitpanda trade file. Skipping."
-                )
-                return
-
-            line = next(reader)
-            assert line in [
-                [
-                    "Order ID",
-                    "Trade ID",
-                    "Type",
-                    "Market",
-                    "Amount",
-                    "Amount Currency",
-                    "Price",
-                    "Price Currency",
-                    "Fee",
-                    "Fee Currency",
-                    "Time (UTC)",
-                ],
-                [
-                    "Order ID",
-                    "Trade ID",
-                    "Type",
-                    "Market",
-                    "Amount",
-                    "Amount Currency",
-                    "Price",
-                    "Price Currency",
-                    "Fee",
-                    "Fee Currency",
-                    "Time (UTC)",
-                    "BEST_EUR Rate",
-                ],
-            ]
-
-            for current_line in reader:
-                if len(current_line) == 11:
-                    (
-                        _order_id,
-                        _trace_id,
-                        operation,
-                        trade_pair,
-                        amount,
-                        amount_currency,
-                        _price,
-                        price_currency,
-                        fee,
-                        fee_currency,
-                        _utc_time,
-                    ) = current_line
-                    best_price = None
-                elif len(current_line) == 12:
-                    (
-                        _order_id,
-                        _trace_id,
-                        operation,
-                        trade_pair,
-                        amount,
-                        amount_currency,
-                        _price,
-                        price_currency,
-                        fee,
-                        fee_currency,
-                        _utc_time,
-                        best_price,
-                    ) = current_line
-                else:
-                    raise NotImplementedError
-
-                row = reader.line_num
-
-                # trade pair is of form e.g. BTC_EUR
-                assert [amount_currency, price_currency] == trade_pair.split("_")
-
-                # At the time of writing (2021-05-02),
-                # there were only these two operations
-                assert operation in ["BUY", "SELL"], "Unsupported operation"
-
-                change = misc.force_decimal(amount)
-                assert change > 0, "Unexpected value for 'Amount' column"
-
-                # see _get_price_bitpanda_pro in price_data.py
-                assert price_currency == "EUR", (
-                    "Only Euro is supported as 'price' currency, "
-                    "since price fetching is not fully implemented yet."
-                )
-
-                # sanity checks
-                assert (
-                    fee_currency == "BEST"
-                    or (operation == "SELL" and fee_currency == price_currency)
-                    or (operation == "BUY" and fee_currency == amount_currency)
-                ), "Invalid fee currency"
-
-                utc_time = misc.parse_iso_timestamp(_utc_time)
-
-                coin = amount_currency
-
-                self.append_operation(
-                    operation.title(), utc_time, platform, change, coin, row, file_path
-                )
-
-                # Save price in our local database for later.
-                price = misc.force_decimal(_price)
-                set_price_db(platform, coin, price_currency, utc_time, price)
-                if best_price:
-                    set_price_db(
-                        platform,
-                        "BEST",
-                        "EUR",
-                        utc_time,
-                        misc.force_decimal(best_price),
-                    )
-
-                self.append_operation(
-                    "Fee",
-                    utc_time,
-                    platform,
-                    misc.force_decimal(fee),
-                    fee_currency,
-                    row,
-                    file_path,
-                )
-
-    def _read_bitpanda(self, file_path: Path) -> None:
-        """Reads a trade statement from Bitpanda.
-
-        Args:
-            file_path (Path): Path to Bitpanda trade history.
-        """
-
-        platform = "bitpanda"
-
-        operation_mapping = {
-            "deposit": "Deposit",
-            "withdrawal": "Withdrawal",
-            "buy": "Buy",
-            "sell": "Sell",
-        }
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-            line = next(reader)
-
-            # skip header, there are multiple lines
-            while line != [
-                "Transaction ID",
-                "Timestamp",
-                "Transaction Type",
-                "In/Out",
-                "Amount Fiat",
-                "Fiat",
-                "Amount Asset",
-                "Asset",
-                "Asset market price",
-                "Asset market price currency",
-                "Asset class",
-                "Product ID",
-                "Fee",
-                "Fee asset",
-                "Spread",
-                "Spread Currency",
-            ]:
-                try:
-                    line = next(reader)
-                except StopIteration:
-                    log.error(f"Expected header not found in file {file_path}")
-                    raise RuntimeError
-
-            for (
-                _tx_id,
-                csv_utc_time,
-                operation,
-                _inout,
-                amount_fiat,
-                fiat,
-                amount_asset,
-                asset,
-                _asset_price,
-                asset_price_currency,
-                asset_class,
-                _product_id,
-                fee,
-                fee_currency,
-                _spread,
-                _spread_currency,
-            ) in reader:
-                row = reader.line_num
-
-                # make RFC3339 timestamp ISO 8601 parseable
-                if csv_utc_time[-1] == "Z":
-                    csv_utc_time = csv_utc_time[:-1] + "+00:00"
-
-                # timezone information is already taken care of with this
-                utc_time = datetime.datetime.fromisoformat(csv_utc_time)
-
-                # transfer ops seem to be akin to airdrops. In my case I got a
-                # CocaCola transfer, which I don't want to track. Would need to
-                # be implemented if need be.
-                if operation == "transfer":
-                    log.warning(
-                        f"'Transfer' operations are not "
-                        f"implemented, skipping row {row} of file {file_path}"
-                    )
-                    continue
-
-                # fail for unknown ops
-                try:
-                    operation = operation_mapping[operation]
-                except KeyError:
-                    log.error(
-                        f"Unsupported operation '{operation}' "
-                        f"in row {row} of file {file_path}"
-                    )
-                    raise RuntimeError
-
-                if operation in ["Deposit", "Withdrawal"]:
-                    if asset_class == "Fiat":
-                        change = misc.force_decimal(amount_fiat)
-                        if fiat != asset:
-                            log.error(
-                                f"Asset {asset} should be {fiat} in "
-                                f"row {row} of file {file_path}"
-                            )
-                            raise RuntimeError
-                    elif asset_class == "Cryptocurrency":
-                        change = misc.force_decimal(amount_asset)
-                    else:
-                        log.error(
-                            f"Unknown asset class {asset_class}: Should be 'Fiat' or "
-                            f"'Cryptocurrency' in row {row} of file {file_path}"
-                        )
-                        raise RuntimeError
-                elif operation in ["Buy", "Sell"]:
-                    if asset_price_currency != config.FIAT:
-                        log.error(
-                            f"Only {config.FIAT} is supported as "
-                            "'Asset market price currency', since price fetching for "
-                            "fiat currencies is not fully implemented yet."
-                        )
-                        raise RuntimeError
-                    change = misc.force_decimal(amount_asset)
-                    change_fiat = misc.force_decimal(amount_fiat)
-                    # Save price in our local database for later.
-                    # Rounded price in CSV
-                    # price = misc.force_decimal(asset_price)
-                    # Calculated price
-                    price_calc = change_fiat / change
-                    set_price_db(platform, asset, config.FIAT, utc_time, price_calc)
-
-                if change < 0:
-                    log.error(
-                        f"Unexpected value for the amount '{change}' of this "
-                        f"{operation} in row {row} of file {file_path}"
-                    )
-                    raise RuntimeError
-
-                self.append_operation(
-                    operation, utc_time, platform, change, asset, row, file_path
-                )
-
-                # add buy / sell operation for fiat currency
-                if operation == "Buy":
-                    self.append_operation(
-                        "Sell",
-                        utc_time,
-                        platform,
-                        change_fiat,
-                        config.FIAT,
-                        row,
-                        file_path,
-                    )
-                elif operation == "Sell":
-                    self.append_operation(
-                        "Buy",
-                        utc_time,
-                        platform,
-                        change_fiat,
-                        config.FIAT,
-                        row,
-                        file_path,
-                    )
-
-                if fee != "-":
-                    self.append_operation(
-                        "Fee",
-                        utc_time,
-                        platform,
-                        misc.force_decimal(fee),
-                        fee_currency,
-                        row,
-                        file_path,
-                    )
-
-    def _read_bitunix(self, file_path: Path) -> None:
-        """Reads a transaction history from Bitunix.
-
-        Args:
-            file_path (Path): Path to Bitunix transaction history CSV.
-        """
-        platform = "bitunix"
-        operation_mapping = {
-            "Deposit": "Deposit",
-            "Withdraw": "Withdrawal",
-            "Spot Trade": "Trade",
-            "Futures Profit": "Profit",
-            "Futures Loss": "Loss",
-            "Rebate (Agent)": "Commission",
-        }
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for columns in reader:
-                if len(columns) != 10:
-                    log.warning(
-                        f"{file_path}: Expected 10 columns, got {len(columns)}. "
-                        "Skipping row."
-                    )
-                    continue
-
-                (
-                    _utc_time,
-                    label,
-                    outgoing_asset,
-                    _outgoing_amount,
-                    incoming_asset,
-                    _incoming_amount,
-                    fee_asset,
-                    _fee_amount,
-                    _trx_id,
-                    _comment,
-                ) = columns
-
-                row = reader.line_num
-
-                # Parse data.
-                utc_time = datetime.datetime.strptime(
-                    _utc_time, "%Y-%m-%d %H:%M:%S"
-                )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-
-                outgoing_amount = misc.xdecimal(_outgoing_amount) or decimal.Decimal(0)
-                incoming_amount = misc.xdecimal(_incoming_amount) or decimal.Decimal(0)
-                fee_amount = misc.xdecimal(_fee_amount) or decimal.Decimal(0)
-
-                operation = operation_mapping.get(label)
-
-                if operation is None:
-                    log.warning(
-                        f"{file_path} row {row}: Unknown operation type '{label}'. "
-                        "Skipping row."
-                    )
-                    continue
-
-                # Handle different operation types
-                if operation == "Deposit":
-                    # Deposit: incoming_asset and incoming_amount
-                    if incoming_amount > 0:
-                        self.append_operation(
-                            "Deposit",
-                            utc_time,
-                            platform,
-                            incoming_amount,
-                            incoming_asset,
-                            row,
-                            file_path,
-                        )
-                    if fee_amount > 0:
-                        self.append_operation(
-                            "Fee",
-                            utc_time,
-                            platform,
-                            fee_amount,
-                            fee_asset,
-                            row,
-                            file_path,
-                        )
-
-                elif operation == "Withdrawal":
-                    # Withdrawal: outgoing_asset and outgoing_amount
-                    if outgoing_amount > 0:
-                        self.append_operation(
-                            "Withdrawal",
-                            utc_time,
-                            platform,
-                            outgoing_amount,
-                            outgoing_asset,
-                            row,
-                            file_path,
-                        )
-                    if fee_amount > 0:
-                        self.append_operation(
-                            "Fee",
-                            utc_time,
-                            platform,
-                            fee_amount,
-                            fee_asset,
-                            row,
-                            file_path,
-                        )
-
-                elif operation == "Trade":
-                    # Spot Trade: sell outgoing, buy incoming
-                    if outgoing_amount > 0 and outgoing_asset:
-                        self.append_operation(
-                            "Sell",
-                            utc_time,
-                            platform,
-                            outgoing_amount,
-                            outgoing_asset,
-                            row,
-                            file_path,
-                        )
-
-                    if incoming_amount > 0 and incoming_asset:
-                        self.append_operation(
-                            "Buy",
-                            utc_time,
-                            platform,
-                            incoming_amount,
-                            incoming_asset,
-                            row,
-                            file_path,
-                        )
-
-                    if fee_amount > 0 and fee_asset:
-                        self.append_operation(
-                            "Fee",
-                            utc_time,
-                            platform,
-                            fee_amount,
-                            fee_asset,
-                            row,
-                            file_path,
-                        )
-
-                elif operation == "Profit":
-                    # Futures Profit: incoming amount is the profit
-                    if incoming_amount > 0 and incoming_asset:
-                        self.append_operation(
-                            "StakingInterest",
-                            utc_time,
-                            platform,
-                            incoming_amount,
-                            incoming_asset,
-                            row,
-                            file_path,
-                        )
-
-                    if fee_amount > 0 and fee_asset:
-                        self.append_operation(
-                            "Fee",
-                            utc_time,
-                            platform,
-                            fee_amount,
-                            fee_asset,
-                            row,
-                            file_path,
-                        )
-
-                elif operation == "Loss":
-                    # Futures Loss: outgoing amount is the loss
-                    # Record as a sell operation (loss is like selling at unfavorable rate)
-                    if outgoing_amount > 0 and outgoing_asset:
-                        self.append_operation(
-                            "Sell",
-                            utc_time,
-                            platform,
-                            outgoing_amount,
-                            outgoing_asset,
-                            row,
-                            file_path,
-                        )
-
-                    if fee_amount > 0 and fee_asset:
-                        self.append_operation(
-                            "Fee",
-                            utc_time,
-                            platform,
-                            fee_amount,
-                            fee_asset,
-                            row,
-                            file_path,
-                        )
-
-                elif operation == "Commission":
-                    # Rebate (Agent): incoming amount is the commission/rebate
-                    if incoming_amount > 0 and incoming_asset:
-                        self.append_operation(
-                            "Commission",
-                            utc_time,
-                            platform,
-                            incoming_amount,
-                            incoming_asset,
-                            row,
-                            file_path,
-                        )
-
-    def _read_pionex(self, file_path: Path) -> None:
-        """Dispatcher method for Pionex CSV files.
-
-        Routes to appropriate reader based on filename.
-
-        Args:
-            file_path (Path): Path to Pionex CSV file
-        """
-        filename = file_path.name
-
-        if filename == "deposit-withdraw.csv":
-            self._read_pionex_deposit_withdraw(file_path)
-        elif filename == "trading.csv":
-            self._read_pionex_trading(file_path)
-        elif filename == "staking.csv":
-            self._read_pionex_staking(file_path)
-        elif filename == "others.csv":
-            self._read_pionex_others(file_path)
-        else:
-            log.warning(
-                f"Unknown Pionex file format: {filename}. "
-                "Skipping file."
-            )
-
-    def _read_pionex_deposit_withdraw(self, file_path: Path) -> None:
-        """Reads deposit/withdrawal records from Pionex.
-
-        Args:
-            file_path (Path): Path to Pionex deposit-withdraw.csv
-        """
-        platform = "pionex"
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for columns in reader:
-                if len(columns) != 7:
-                    log.warning(
-                        f"{file_path}: Expected 7 columns, got {len(columns)}. "
-                        "Skipping row."
-                    )
-                    continue
-
-                (
-                    _utc_time,
-                    tx_type,
-                    _amount,
-                    coin,
-                    _network,
-                    _txid,
-                    _fee,
-                ) = columns
-
-                row = reader.line_num
-
-                # Parse data.
-                utc_time = datetime.datetime.strptime(
-                    _utc_time, "%Y-%m-%d %H:%M:%S"
-                )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-
-                amount = misc.force_decimal(_amount)
-                fee = misc.xdecimal(_fee) or decimal.Decimal(0)
-
-                # Map transaction type
-                if tx_type == "DEPOSIT":
-                    self.append_operation(
-                        "Deposit",
-                        utc_time,
-                        platform,
-                        amount,
-                        coin,
-                        row,
-                        file_path,
-                    )
-                elif tx_type == "WITHDRAW":
-                    self.append_operation(
-                        "Withdrawal",
-                        utc_time,
-                        platform,
-                        amount,
-                        coin,
-                        row,
-                        file_path,
-                    )
-                    # Add withdrawal fee if present
-                    if fee > 0:
-                        self.append_operation(
-                            "Fee",
-                            utc_time,
-                            platform,
-                            fee,
-                            coin,
-                            row,
-                            file_path,
-                        )
-                else:
-                    log.warning(
-                        f"{file_path} row {row}: Unknown tx_type '{tx_type}'. "
-                        "Skipping row."
-                    )
-
-    def _read_pionex_trading(self, file_path: Path) -> None:
-        """Reads trading records from Pionex (spot and futures).
-
-        Args:
-            file_path (Path): Path to Pionex trading.csv
-        """
-        platform = "pionex"
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for columns in reader:
-                if len(columns) != 10:
-                    log.warning(
-                        f"{file_path}: Expected 10 columns, got {len(columns)}. "
-                        "Skipping row."
-                    )
-                    continue
-
-                (
-                    _utc_time,
-                    _executed_qty,
-                    _amount,
-                    _price,
-                    side,
-                    symbol,
-                    _fee,
-                    fee_coin,
-                    _market_type,
-                    _tax_id,
-                ) = columns
-
-                row = reader.line_num
-
-                # Parse data.
-                utc_time = datetime.datetime.strptime(
-                    _utc_time, "%Y-%m-%d %H:%M:%S"
-                )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-
-                executed_qty = misc.force_decimal(_executed_qty)
-                amount = misc.force_decimal(_amount)
-                fee = misc.force_decimal(_fee)
-
-                # Extract base and quote coin from symbol (e.g. DOT_USDT_PERP -> DOT, USDT)
-                # Format: BASE_QUOTE or BASE_QUOTE_PERP
-                symbol_parts = symbol.split("_")
-                if len(symbol_parts) >= 2:
-                    base_coin = symbol_parts[0]
-                    quote_coin = symbol_parts[1]
-                else:
-                    log.warning(
-                        f"{file_path} row {row}: Could not parse symbol '{symbol}'. "
-                        "Skipping row."
-                    )
-                    continue
-
-                # Record buy/sell transactions
-                if side == "BUY":
-                    # Buying: spending quote coin, getting base coin
-                    self.append_operation(
-                        "Buy",
-                        utc_time,
-                        platform,
-                        executed_qty,
-                        base_coin,
-                        row,
-                        file_path,
-                    )
-                    self.append_operation(
-                        "Sell",
-                        utc_time,
-                        platform,
-                        amount,
-                        quote_coin,
-                        row,
-                        file_path,
-                    )
-                elif side == "SELL":
-                    # Selling: spending base coin, getting quote coin
-                    self.append_operation(
-                        "Sell",
-                        utc_time,
-                        platform,
-                        executed_qty,
-                        base_coin,
-                        row,
-                        file_path,
-                    )
-                    self.append_operation(
-                        "Buy",
-                        utc_time,
-                        platform,
-                        amount,
-                        quote_coin,
-                        row,
-                        file_path,
-                    )
-                else:
-                    log.warning(
-                        f"{file_path} row {row}: Unknown side '{side}'. "
-                        "Skipping row."
-                    )
-                    continue
-
-                # Add trading fee if present
-                if fee > 0:
-                    self.append_operation(
-                        "Fee",
-                        utc_time,
-                        platform,
-                        fee,
-                        fee_coin,
-                        row,
-                        file_path,
-                    )
-
-    def _read_pionex_staking(self, file_path: Path) -> None:
-        """Reads staking records from Pionex.
-
-        Args:
-            file_path (Path): Path to Pionex staking.csv
-        """
-        platform = "pionex"
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for columns in reader:
-                if len(columns) != 6:
-                    log.warning(
-                        f"{file_path}: Expected 6 columns, got {len(columns)}. "
-                        "Skipping row."
-                    )
-                    continue
-
-                (
-                    _utc_time,
-                    _received_qty,
-                    received_currency,
-                    _sent_qty,
-                    sent_currency,
-                    _tag,
-                ) = columns
-
-                row = reader.line_num
-
-                # Skip empty rows
-                if not _received_qty.strip() or _received_qty == "0":
-                    continue
-
-                # Parse data.
-                utc_time = datetime.datetime.strptime(
-                    _utc_time, "%Y-%m-%d %H:%M:%S"
-                )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-
-                received_qty = misc.xdecimal(_received_qty) or decimal.Decimal(0)
-                sent_qty = misc.xdecimal(_sent_qty) or decimal.Decimal(0)
-
-                # Staking record: sent asset (being staked) and received interest
-                if sent_qty > 0 and sent_currency:
-                    self.append_operation(
-                        "Staking",
-                        utc_time,
-                        platform,
-                        sent_qty,
-                        sent_currency,
-                        row,
-                        file_path,
-                    )
-
-                if received_qty > 0 and received_currency:
-                    self.append_operation(
-                        "StakingInterest",
-                        utc_time,
-                        platform,
-                        received_qty,
-                        received_currency,
-                        row,
-                        file_path,
-                    )
-
-    def _read_pionex_others(self, file_path: Path) -> None:
-        """Reads other transaction records from Pionex (fees, funding, etc).
-
-        Args:
-            file_path (Path): Path to Pionex others.csv
-        """
-        platform = "pionex"
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for columns in reader:
-                if len(columns) != 5:
-                    log.warning(
-                        f"{file_path}: Expected 5 columns, got {len(columns)}. "
-                        "Skipping row."
-                    )
-                    continue
-
-                (
-                    _utc_time,
-                    coin,
-                    _amount,
-                    tag,
-                    _comment,
-                ) = columns
-
-                row = reader.line_num
-
-                # Parse data.
-                utc_time = datetime.datetime.strptime(
-                    _utc_time, "%Y-%m-%d %H:%M:%S"
-                )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-
-                amount = misc.force_decimal(_amount)
-                amount_abs = abs(amount)
-
-                # Handle different tag types
-                if tag == "FundingFee":
-                    # Funding fees are costs, record as Fee
-                    if amount_abs > 0:
-                        self.append_operation(
-                            "Fee",
-                            utc_time,
-                            platform,
-                            amount_abs,
-                            coin,
-                            row,
-                            file_path,
-                        )
-                elif tag == "Commission":
-                    # Commission/rewards
-                    if amount > 0:
-                        self.append_operation(
-                            "Commission",
-                            utc_time,
-                            platform,
-                            amount,
-                            coin,
-                            row,
-                            file_path,
-                        )
-                elif tag == "Airdrop":
-                    # Airdrop/reward distribution
-                    if amount > 0:
-                        self.append_operation(
-                            "Airdrop",
-                            utc_time,
-                            platform,
-                            amount,
-                            coin,
-                            row,
-                            file_path,
-                        )
-                else:
-                    # Log unknown tags for reference
-                    if amount > 0:
-                        log.debug(
-                            f"{file_path} row {row}: Unknown tag '{tag}' "
-                            f"with amount {amount}. Recording as income."
-                        )
-                        self.append_operation(
-                            "Airdrop",
-                            utc_time,
-                            platform,
-                            amount,
-                            coin,
-                            row,
-                            file_path,
-                        )
-
-    def _read_custom_eur(self, file_path: Path) -> None:
-        fiat = "EUR"
-
-        with open(file_path, encoding="utf8") as f:
-            reader = csv.reader(f)
-
-            # Skip header.
-            next(reader)
-
-            for line in reader:
-                row = reader.line_num
-
-                # Skip empty lines.
-                if not line:
-                    continue
-
-                (
-                    operation_type,
-                    _buy_quantity,
-                    buy_asset,
-                    _buy_value_in_fiat,
-                    _sell_quantity,
-                    sell_asset,
-                    _sell_value_in_fiat,
-                    _fee_quantity,
-                    fee_asset,
-                    _fee_value_in_fiat,
-                    platform,
-                    _timestamp,
-                    remark,
-                ) = line
-
-                # Parse data.
-                try:
-                    utc_time = datetime.datetime.strptime(
-                        _timestamp, "%m/%d/%Y %H:%M:%S"
-                    )
-                except ValueError:
-                    utc_time = datetime.datetime.strptime(
-                        _timestamp, "%m/%d/%Y %H:%M:%S.%f"
-                    )
-                utc_time = utc_time.replace(tzinfo=datetime.timezone.utc)
-                buy_quantity = misc.xdecimal(_buy_quantity)
-                buy_value_in_fiat = misc.xdecimal(_buy_value_in_fiat)
-                sell_quantity = misc.xdecimal(_sell_quantity)
-                sell_value_in_fiat = misc.xdecimal(_sell_value_in_fiat)
-                fee_quantity = misc.xdecimal(_fee_quantity)
-                fee_value_in_fiat = misc.xdecimal(_fee_value_in_fiat)
-
-                # ... and define which operation to add.
-                add_operations: list[
-                    tuple[str, decimal.Decimal, str, Optional[decimal.Decimal]]
-                ] = []
-                if operation_type != "Withdrawal":
-                    assert buy_asset
-
-                if operation_type not in ("Deposit", "Airdrop"):
-                    assert sell_asset
-
-                if fee_asset:
-                    assert fee_quantity
-
-                # Map operation_type to standard operation
-                operation = operation_type  # or use a mapping if needed
-
-                if buy_quantity and buy_quantity != 0:
-                    add_operations.append(
-                        (operation, buy_quantity, buy_asset, buy_value_in_fiat))
-
-                if sell_quantity and sell_quantity != 0:
-                    # For sell, quantity is negative
-                    add_operations.append(
-                        (operation, -sell_quantity, sell_asset, sell_value_in_fiat))
-
-                if fee_quantity and fee_quantity != 0:
-                    add_operations.append(
-                        ("Fee", -fee_quantity, fee_asset, fee_value_in_fiat))
-
-                for operation, change, coin, change_in_fiat in add_operations:
-                    assert change
-                    assert coin
-                    self.append_operation(
-                        operation,
-                        utc_time,
-                        platform,
-                        change,
-                        coin,
-                        row,
-                        file_path,
-                        remark=remark,
-                    )
-                    # Add price from csv.
-                    if change_in_fiat and coin != fiat:
-                        price = change_in_fiat / change
-                        log.debug(
-                            f"Adding {fiat}/{coin} price from custom CSV: "
-                            f"{price} for {platform} at {utc_time}"
-                        )
-                        set_price_db(
-                            platform,
-                            coin,
-                            fiat,
-                            utc_time,
-                            price,
-                            overwrite=True,
-                        )
-
-    def _bitget_headers(self, method: str, request_path: str, body: str = "") -> dict[str, str]:
-        if not (
-            config.BITGET_API_KEY
-            and config.BITGET_API_SECRET
-            and config.BITGET_API_PASSPHRASE
-        ):
-            raise RuntimeError(
-                "Bitget API credentials are not configured. Set BITGET_API_KEY, "
-                "BITGET_API_SECRET and BITGET_API_PASSPHRASE."
-            )
-
-        timestamp = str(int(time.time() * 1000))
-        signature_payload = f"{timestamp}{method.upper()}{request_path}{body}"
-        signature = base64.b64encode(
-            hmac.new(
-                config.BITGET_API_SECRET.encode(),
-                signature_payload.encode(),
-                hashlib.sha256,
-            ).digest()
-        ).decode()
-
-        return {
-            "ACCESS-KEY": config.BITGET_API_KEY,
-            "ACCESS-SIGN": signature,
-            "ACCESS-TIMESTAMP": timestamp,
-            "ACCESS-PASSPHRASE": config.BITGET_API_PASSPHRASE,
-            "Content-Type": "application/json",
-            "locale": "en-US",
-        }
-
-    def _bitget_get(
-        self, path: str, params: Optional[dict[str, Any]] = None
-    ) -> list[dict[str, Any]]:
-        params = params or {}
-        params = {str(k): str(v) for k, v in params.items()}
-        query_string = f"?{urlencode(sorted(params.items()))}" if params else ""
-        request_path = f"{path}{query_string}"
-        url = f"{config.BITGET_API_BASE_URL}{request_path}"
-
-        log.debug("Calling Bitget API %s", url)
-        try:
-            response = requests.get(
-                url, headers=self._bitget_headers("GET", request_path))
-            response.raise_for_status()
-            data = response.json()
-
-        except requests.exceptions.HTTPError as e:
-            print("HTTP Error:", e)
-            print("Status Code:", response.status_code)
-            print("Response Text:", response.text)
-
-            # Falls Bitget JSON zurückgibt:
-            try:
-                print("JSON Error:", response.json())
-            except ValueError:
-                print("No JSON body")
-
-            raise  # optional: Fehler weiterwerfen
-
-        time.sleep(1)
-
-        if not isinstance(data, dict):
-            raise RuntimeError("Bitget API returned unexpected response format.")
-
-        if data.get("code") != "00000":
-            raise RuntimeError(
-                f"Bitget API returned error {data.get('code')} {data.get('msg')}"
-            )
-
-        return data.get("data", [])
-
-    def _bitget_fetch_all(
-        self, path: str, params: dict[str, Any]
-    ) -> list[dict[str, Any]]:
-        records: list[dict[str, Any]] = []
-        params = params.copy()
-        params.setdefault("limit", 500)
-
-        while True:
-            page = self._bitget_get(path, params)
-            if not page:
-                break
-            records.extend(page)
-            if len(page) < int(params["limit"]):
-                break
-            params["idLessThan"] = page[-1]["id"]
-
-        return records
-
-    def _get_bitget_resume_state_path(self) -> Path:
-        resume_dir = Path(config.DATA_PATH)
-        resume_dir.mkdir(parents=True, exist_ok=True)
-        return resume_dir / "bitget_resume_state.json"
-
-    def _load_bitget_resume_state(self) -> dict[str, Any]:
-        path = self._get_bitget_resume_state_path()
-        if not path.exists():
-            return {}
-        try:
-            return json.loads(path.read_text(encoding="utf-8"))
-        except ValueError:
-            log.warning("Unable to read existing Bitget resume state, resetting it.")
-            return {}
-
-    def _save_bitget_resume_state(self, state: dict[str, Any]) -> None:
-        path = self._get_bitget_resume_state_path()
-        path.write_text(json.dumps(state, indent=2), encoding="utf-8")
-
-    def _get_bitget_cached_chunk(
-        self, endpoint_state: dict[str, Any], start_ms: int, end_ms: int
-    ) -> Optional[dict[str, Any]]:
-        for chunk in endpoint_state.setdefault("chunks", []):
-            if (
-                chunk.get("start_ms") == start_ms
-                and chunk.get("end_ms") == end_ms
-            ):
-                return chunk
-        return None
-
-    def _mark_bitget_chunk_processed(
-        self,
-        resume_state: dict[str, Any],
-        endpoint_state: dict[str, Any],
-        start_ms: int,
-        end_ms: int,
-    ) -> None:
-        for chunk in endpoint_state.setdefault("chunks", []):
-            if chunk.get("start_ms") == start_ms and chunk.get("end_ms") == end_ms:
-                chunk["processed"] = True
-                break
-        endpoint_state["last_end_ms"] = end_ms
-        endpoint_state["updated_at"] = int(time.time() * 1000)
-        self._save_bitget_resume_state(resume_state)
-
-    def _bitget_fetch_all_range(
-        self,
-        path: str,
-        start_time_ms: int,
-        end_time_ms: int,
-        extra_params: Optional[dict[str, Any]] = None,
-    ) -> list[tuple[int, int, list[dict[str, Any]], dict[str, Any], dict[str, Any]]]:
-        cursor_start = start_time_ms
-        chunk_size = 10 * 24 * 60 * 60 * 1000
-        endpoint_key = path.strip("/").replace("/", "_")
-
-        resume_state = self._load_bitget_resume_state()
-        if resume_state.get("tax_year") != config.TAX_YEAR:
-            resume_state = {"tax_year": config.TAX_YEAR, "endpoints": {}}
-
-        endpoints = resume_state.setdefault("endpoints", {})
-        endpoint_state = endpoints.setdefault(endpoint_key, {})
-        endpoint_state.setdefault("chunks", [])
-        last_end_ms = endpoint_state.get("last_end_ms", start_time_ms - 1)
-
-        chunks: list[tuple[int, int, list[dict[str, Any]],
-                           dict[str, Any], dict[str, Any]]] = []
-        while cursor_start <= end_time_ms:
-            cursor_end = min(cursor_start + chunk_size - 1, end_time_ms)
-            params = {
-                "startTime": cursor_start,
-                "endTime": cursor_end,
-                "limit": 500,
-            }
-            if extra_params:
-                params.update(extra_params)
-
-            cached_chunk = self._get_bitget_cached_chunk(
-                endpoint_state, cursor_start, cursor_end
-            )
-            if cached_chunk is not None:
-                log.info(
-                    "Reusing cached Bitget API records %s - %s",
-                    datetime.datetime.fromtimestamp(
-                        cursor_start / 1000, tz=datetime.timezone.utc),
-                    datetime.datetime.fromtimestamp(
-                        cursor_end / 1000, tz=datetime.timezone.utc),
-                )
-                segment = cached_chunk["records"]
-            else:
-                log.info(
-                    "Fetching Bitget API records %s - %s",
-                    datetime.datetime.fromtimestamp(
-                        cursor_start / 1000, tz=datetime.timezone.utc),
-                    datetime.datetime.fromtimestamp(
-                        cursor_end / 1000, tz=datetime.timezone.utc),
-                )
-                segment = self._bitget_fetch_all(path, params)
-                endpoint_state["chunks"].append(
-                    {
-                        "start_ms": cursor_start,
-                        "end_ms": cursor_end,
-                        "records": segment,
-                        "processed": False,
-                    }
-                )
-                self._save_bitget_resume_state(resume_state)
-
-            chunks.append(
-                (cursor_start, cursor_end, segment, resume_state, endpoint_state)
-            )
-
-            if cursor_end == end_time_ms:
-                break
-            cursor_start = cursor_end + 1
-
-        return chunks
-
-    def _map_bitget_spot_tax_type(self, spot_tax_type: str) -> Optional[str]:
-        normalized = spot_tax_type.strip()
-        mapping = {
-            "Deposit": "Deposit",
-            "Withdrawal": "Withdrawal",
-            "Buy": "Buy",
-            "Sell": "Sell",
-            "Buy Crypto": "Buy",
-            "Sell Crypto": "Sell",
-            "Crypto Voucher Distribution": "Airdrop",
-            "Airdrop Reward-A": "Airdrop",
-            "Airdrop Reward-B": "Airdrop",
-            "Interest": "StakingInterest",
-            "batch_interest_user_in": "StakingInterest",
-            "User fees": "Fee",
-            "Transaction fee deduct": "Fee",
-            "System charges fees": "Fee",
-            "financial_lock_out": "Fee",
-            "Trading fee rebate": "Commission",
-            "Fiat withdrawal success - Deduct": "Withdrawal",
-            "Reward": "Airdrop",
-            "Automatic deposit": "Deposit",
-            "Automatic withdrawal": "Withdrawal",
-            "Transfer in": "Deposit",
-            "Transfer out": "Withdrawal",
-            "fiat_recharge_in": "Deposit",
-            "fiat_balance_success_user_in": "Deposit",
-            "fiat_balance_user_out": "Withdrawal",
-            "financial_rede_in": "Deposit",
-            "financial_unlock_in": "Deposit",
-            "financial_pos_out": "Withdrawal",
-            "financial_subs_out": "Withdrawal",
-            "Copy Trade expense": "Fee",
-            "Refund Copy Trade commission": "Commission",
-            "Consumption": "Sell",
-            "Gains": "Airdrop",
-        }
-
-        if normalized in mapping:
-            return mapping[normalized]
-
-        lower_name = normalized.lower()
-        if lower_name.endswith("_in"):
-            return "Deposit"
-        if lower_name.endswith("_out"):
-            return "Withdrawal"
-        if "commission" in lower_name:
-            return "Commission"
-        if "expense" in lower_name or "fee" in lower_name:
-            return "Fee"
-
-        return mapping.get(normalized)
-
-    def _map_bitget_future_tax_type(self, future_tax_type: str) -> Optional[str]:
-        future_tax_type = future_tax_type.upper()
-        mapping = {
-            "TRANSFER_IN": "Deposit",
-            "TRANSFER_OUT": "Withdrawal",
-            "ORDER_DEALT_IN": "Deposit",
-            "ORDER_DEALT_FROZEN_OUT": "Fee",
-            "ORDER_PLF_FEE_OUT": "Fee",
-            "EXCHANGE_SOURCE_TOKEN_USER_OUT": "Sell",
-            "EXCHANGE_TARGET_TOKEN_USER_IN": "Buy",
-            "OPEN_LONG": "Buy",
-            "OPEN_SHORT": "Buy",
-            "CLOSE_LONG": "Sell",
-            "CLOSE_SHORT": "Sell",
-            "BUY_DEAL": "Buy",
-            "SELL_DEAL": "Sell",
-            "FORCE_CLOSE_LONG": "Sell",
-            "FORCE_CLOSE_SHORT": "Sell",
-            "BURST_CLOSE_LONG": "Sell",
-            "BURST_CLOSE_SHORT": "Sell",
-            "INTEREST_SETTLEMENT_OUT": "Fee",
-            "CONTRACT_MAIN_SETTLE_FEE_USER_IN": "Commission",
-            "CONTRACT_MAIN_SETTLE_FEE_USER_OUT": "Fee",
-            "BONUS_ISSUE": "Airdrop",
-            "RISK_CAPTITAL_USER_TRANSFER": "Deposit",
-        }
-
-        if result := mapping.get(future_tax_type):
-            return result
-
-        if future_tax_type.startswith("TRANS_FROM_"):
-            return "Deposit"
-        if future_tax_type.startswith("TRANS_TO_"):
-            return "Withdrawal"
-        if future_tax_type.startswith("TRANSFER_FROM_"):
-            return "Deposit"
-        if future_tax_type.startswith("TRANSFER_TO_"):
-            return "Withdrawal"
-        if future_tax_type.endswith("_FEE") or "SETTLE_FEE" in future_tax_type:
-            return "Fee"
-        if future_tax_type.endswith("_LONG"):
-            return "Buy" if "OPEN" in future_tax_type else "Sell"
-        if future_tax_type.endswith("_SHORT"):
-            return "Buy" if "OPEN" in future_tax_type else "Sell"
-        return None
-
-    def _map_bitget_margin_tax_type(self, margin_tax_type: str) -> Optional[str]:
-        mapping = {
-            "transfer_in": "Deposit",
-            "transfer_out": "Withdrawal",
-            "borrow": "Deposit",
-            "repay": "Withdrawal",
-            "liquidation_fee": "Fee",
-            "deal_in": "Buy",
-            "deal_out": "Sell",
-            "interest_repay": "Fee",
-            "exchange_in": "Buy",
-            "exchange_out": "Sell",
-            "confiscated": "Fee",
-            "compensate": "Fee",
-        }
-        return mapping.get(margin_tax_type)
-
-    def _map_bitget_p2p_tax_type(self, p2p_tax_type: str) -> Optional[str]:
-        mapping = {
-            "BUY": "Buy",
-            "SELL": "Sell",
-            "TRANSFER_IN": "Deposit",
-            "TRANSFER_OUT": "Withdrawal",
-            "FEE": "Fee",
-            "SERVICE_FEE": "Fee",
-            "REFUND": "Buy",
-            "COMMISSION": "Commission",
-        }
-        return mapping.get(p2p_tax_type)
-
-    def _import_bitget_p2p_records(
-        self, start_time_ms: int, end_time_ms: int
-    ) -> None:
-        chunks = self._bitget_fetch_all_range(
-            "/api/v2/tax/p2p-record",
-            start_time_ms,
-            end_time_ms,
-        )
-
-        if not chunks:
-            log.info("No Bitget P2P account records were returned.")
-            return
-
-        total_records = sum(len(segment) for _, _, segment, _, _ in chunks)
-        log.info("Importing %s Bitget P2P records.", total_records)
-        for chunk_start, chunk_end, records, resume_state, endpoint_state in chunks:
-            for row_num, row in enumerate(records, start=1):
-                tax_type = row.get("p2pTaxType", row.get("taxType", ""))
-                operation = self._map_bitget_p2p_tax_type(tax_type)
-                if operation is None:
-                    log.warning(
-                        "Unsupported Bitget P2P tax type '%s', skipping row %s.",
-                        tax_type,
-                        row_num,
-                    )
-                    continue
-
-                coin = (
-                    row.get("coin")
-                    or row.get("amountCoin")
-                    or row.get("orderCoin")
-                    or "UNKNOWN"
-                )
-                change = abs(
-                    misc.force_decimal(
-                        row.get("amount", row.get("quantity", "0"))
-                    )
-                )
-                fee = abs(misc.force_decimal(row.get("fee", "0")))
-                utc_time = datetime.datetime.fromtimestamp(
-                    int(row.get("ts", row.get("createTime", 0))) / 1000.0,
-                    datetime.timezone.utc,
-                )
-                remark = (
-                    f"Bitget P2P record {row.get('bizOrderId', '')}"
-                    if row.get("bizOrderId")
-                    else f"Bitget P2P record {row.get('orderId', '')}"
-                )
-                self.append_operation(
-                    operation,
-                    utc_time,
-                    "bitget",
-                    change,
-                    coin,
-                    row_num,
-                    Path("bitget-api"),
-                    remark=remark,
-                )
-                if fee and operation not in ("Fee", "Commission"):
-                    self.append_operation(
-                        "Fee",
-                        utc_time,
-                        "bitget",
-                        fee,
-                        coin,
-                        row_num,
-                        Path("bitget-api"),
-                        remark=f"Bitget P2P fee for {tax_type}",
-                    )
-
-    def _import_bitget_spot_records(
-        self, start_time_ms: int, end_time_ms: int
-    ) -> None:
-        chunks = self._bitget_fetch_all_range(
-            "/api/v2/tax/spot-record",
-            start_time_ms,
-            end_time_ms,
-        )
-
-        if not chunks:
-            log.info("No Bitget spot account records were returned.")
-            return
-
-        total_records = sum(len(segment) for _, _, segment, _, _ in chunks)
-        log.info("Importing %s Bitget spot records.", total_records)
-        for chunk_start, chunk_end, records, resume_state, endpoint_state in chunks:
-            for row_num, row in enumerate(records, start=1):
-                tax_type = row.get("spotTaxType", "")
-                operation = self._map_bitget_spot_tax_type(tax_type)
-                if operation is None:
-                    log.warning(
-                        "Unsupported Bitget spot tax type '%s', skipping row %s.",
-                        tax_type,
-                        row_num,
-                    )
-                    continue
-
-                coin = row.get("coin") or "UNKNOWN"
-                change = abs(misc.force_decimal(row.get("amount", "0")))
-                fee = abs(misc.force_decimal(row.get("fee", "0")))
-                utc_time = datetime.datetime.fromtimestamp(
-                    int(row.get("ts", 0)) / 1000.0,
-                    datetime.timezone.utc,
-                )
-                remark = f"Bitget spot record {row.get('bizOrderId', '')}"
-                self.append_operation(
-                    operation,
-                    utc_time,
-                    "bitget",
-                    change,
-                    coin,
-                    row_num,
-                    Path("bitget-api"),
-                    remark=remark,
-                )
-                if fee and operation not in ("Fee", "Commission"):
-                    self.append_operation(
-                        "Fee",
-                        utc_time,
-                        "bitget",
-                        fee,
-                        coin,
-                        row_num,
-                        Path("bitget-api"),
-                        remark=f"Bitget spot fee for {tax_type}",
-                    )
-
-    def _import_bitget_future_records(
-        self, start_time_ms: int, end_time_ms: int
-    ) -> None:
-        chunks = self._bitget_fetch_all_range(
-            "/api/v2/tax/future-record",
-            start_time_ms,
-            end_time_ms,
-        )
-
-        if not chunks:
-            log.info("No Bitget future account records were returned.")
-            return
-
-        total_records = sum(len(segment) for _, _, segment, _, _ in chunks)
-        log.info("Importing %s Bitget future records.", total_records)
-        for chunk_start, chunk_end, records, resume_state, endpoint_state in chunks:
-            for row_num, row in enumerate(records, start=1):
-                tax_type = row.get("futureTaxType", "")
-                operation = self._map_bitget_future_tax_type(tax_type)
-                if operation is None:
-                    log.warning(
-                        "Unsupported Bitget future tax type '%s', skipping row %s.",
-                        tax_type,
-                        row_num,
-                    )
-                    continue
-
-                coin = row.get("marginCoin") or row.get(
-                    "symbol") or row.get("coin") or "UNKNOWN"
-                change = abs(misc.force_decimal(row.get("amount", "0")))
-                fee = abs(misc.force_decimal(row.get("fee", "0")))
-                utc_time = datetime.datetime.fromtimestamp(
-                    int(row.get("ts", 0)) / 1000.0,
-                    datetime.timezone.utc,
-                )
-                remark = f"Bitget future record {row.get('symbol', '')}"
-                self.append_operation(
-                    operation,
-                    utc_time,
-                    "bitget",
-                    change,
-                    coin,
-                    row_num,
-                    Path("bitget-api"),
-                    remark=remark,
-                )
-                if fee and operation not in ("Fee", "Commission"):
-                    self.append_operation(
-                        "Fee",
-                        utc_time,
-                        "bitget",
-                        fee,
-                        coin,
-                        row_num,
-                        Path("bitget-api"),
-                        remark=f"Bitget future fee for {tax_type}",
-                    )
-
-    def _import_bitget_margin_records(
-        self, start_time_ms: int, end_time_ms: int
-    ) -> None:
-        chunks = self._bitget_fetch_all_range(
-            "/api/v2/tax/margin-record",
-            start_time_ms,
-            end_time_ms,
-        )
-
-        if not chunks:
-            log.info("No Bitget margin account records were returned.")
-            return
-
-        total_records = sum(len(segment) for _, _, segment, _, _ in chunks)
-        log.info("Importing %s Bitget margin records.", total_records)
-        for chunk_start, chunk_end, records, resume_state, endpoint_state in chunks:
-            for row_num, row in enumerate(records, start=1):
-                tax_type = row.get("marginTaxType", "")
-                operation = self._map_bitget_margin_tax_type(tax_type)
-                if operation is None:
-                    log.warning(
-                        "Unsupported Bitget margin tax type '%s', skipping row %s.",
-                        tax_type,
-                        row_num,
-                    )
-                    continue
-
-                coin = row.get("coin") or row.get("symbol") or "UNKNOWN"
-                change = abs(misc.force_decimal(row.get("amount", "0")))
-                fee = abs(misc.force_decimal(row.get("fee", "0")))
-                utc_time = datetime.datetime.fromtimestamp(
-                    int(row.get("ts", 0)) / 1000.0,
-                    datetime.timezone.utc,
-                )
-                remark = f"Bitget margin record {row.get('symbol', '')}"
-                self.append_operation(
-                    operation,
-                    utc_time,
-                    "bitget",
-                    change,
-                    coin,
-                    row_num,
-                    Path("bitget-api"),
-                    remark=remark,
-                )
-                if fee and operation not in ("Fee", "Commission"):
-                    self.append_operation(
-                        "Fee",
-                        utc_time,
-                        "bitget",
-                        fee,
-                        coin,
-                        row_num,
-                        Path("bitget-api"),
-                        remark=f"Bitget margin fee for {tax_type}",
-                    )
-
     def import_bitget_api_records(self) -> None:
         log.info("Importing Bitget records from API for tax year %s.", config.TAX_YEAR)
         year_start = datetime.datetime(
@@ -2580,20 +184,21 @@ class Book:
         year_end = datetime.datetime(
             config.TAX_YEAR, 12, 31, 23, 59, 59, 999000, tzinfo=datetime.timezone.utc
         )
-        self._import_bitget_spot_records(
-            int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
+        bitget_reader = BitgetApiReader()
+        bitget_reader.import_spot_records(
+            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
         )
-        self._import_bitget_future_records(
-            int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
+        bitget_reader.import_future_records(
+            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
         )
-        self._import_bitget_margin_records(
-            int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
+        bitget_reader.import_margin_records(
+            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
         )
-        self._import_bitget_p2p_records(
-            int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
+        bitget_reader.import_p2p_records(
+            self, int(year_start.timestamp() * 1000), int(year_end.timestamp() * 1000)
         )
 
-    def detect_exchange(self, file_path: Path) -> Optional[str]:
+    def detect_exchange(self, file_path: Path) -> Optional[ExchangeReader]:
         if file_path.suffix == ".csv":
 
             expected_header_row = {
@@ -2841,7 +446,7 @@ class Book:
                     expected = expected_headers[exchange_type]
                     header = next(reader, None)
                     if header == expected:
-                        return exchange_type
+                        return create_exchange_reader(exchange_type)
 
             with open(file_path, encoding="utf8") as f:
                 reader = csv.reader(f)
@@ -2855,7 +460,7 @@ class Book:
                     for _ in range(header_row_num):
                         header = next(reader, None)
                         if header == expected:
-                            return exchange
+                            return create_exchange_reader(exchange)
                     # rewind the file after each header check
                     f.seek(0)
 
@@ -3205,19 +810,9 @@ class Book:
         """
         assert file_path.is_file()
 
-        if exchange := self.detect_exchange(file_path):
-
-            try:
-                read_file = getattr(self, f"_read_{exchange}")
-            except AttributeError:
-                log.warning(
-                    f"Unable to read files from the exchange `{exchange}`. "
-                    f"Skipping `{file_path}`."
-                )
-                return
-
-            log.info("Reading file from exchange %s at %s", exchange, file_path)
-            read_file(file_path)
+        if reader := self.detect_exchange(file_path):
+            log.info("Reading file from exchange %s at %s", reader.platform, file_path)
+            reader.read_file(file_path, self)
         elif file_path.suffix not in (
             ".zip",
             ".rar",
